requirements.txt
streamlit
yfinance
pandas
numpy
plotly
tensorflow
scikit-learn
pandas-ta
pillow
openai
